+++ frontend/src/api/chat.js
@@ -36,7 +36,7 @@ async function parseSSEStream(response, callbacks) {
           else if (evt.type === 'error') onError?.(evt.error);
           else if (evt.type === 'title_updated') onTitleUpdated?.(evt.title);
           else if (evt.type === 'memory_recall_start') onMemoryRecallStart?.();
-          else if (evt.type === 'memory_recall_done') onMemoryRecallDone?.();
+          else if (evt.type === 'memory_recall_done') onMemoryRecallDone?.(evt);
           else if (evt.type === 'memory_expand_start') onMemoryExpandStart?.(evt);
           else if (evt.type === 'memory_expand_done') onMemoryExpandDone?.(evt);
         } catch {